Added normalizer class for normalization of expression values
Refactoring of qtable expression value normalization with the qanalysis
module. Fitting for normalization and transformation of values is now
done by normalizer classes from the normalize module. In the qanalysis
module now one single function is used for normalization of qtables,
normalize_expression(). normalize_expression() returns a normalizer
instance that could be used for subsequently applying the same
normalization to the same samples or to look up / plot the fits used
for normalization.