Added missingness analysis, which directly adds columns to the qtable
analyze_missingness() addes columns "Missing ..." and "Events ...",
which count the number of missing values or the number of
quantification events, respectively. The functions validate_proteins()
from the qanalysis module, and the plotting functions for missing
values now rely on the columns added by analyze_missingness(), and
don't calculate the missing values themselves anymore.