Fix average expression calculation of calculate_multi_group_limma()
When multiple experiments are present, the average expression calculated
by limma is always the average of all samples, independent which contrast
is analysed. However, the average expression should actually correspond
to the average intensity of the compared experiments. Hence logFC and
average expression are now re-calculated using the two_group_comparison().