feat!: Change interface of normalize_expression() to require a Normalizer instance
Breaking change to the interface of msreport.analyze.normalize_expression(),
which now requires passing a Normalizer instance from the msreport.normalize module
(or any class that complies with the msreport.analyze.Transformer protocol).
This update makes it possible to add new normalization strategies by creating
additional Normalizer classes, without affecting the way that normalization of
missing expression values is done for a qtable.