fix: Issue with `config` argument when creating protein excel reports
It was only checked if the specified `config` argument corresponds to
a file present in the xlsxreport appdata directory and not if it is
a valid file path. The new behaviour is to first check if the file
exists, and if not, to check if `config` corresponds to a file in the
xlsxreport appdata directory. If neither is true, then a file not
found error is raised.